Limiting worldname length to 16 characters
Limiting worldname length to 16 characters.  This is the maximum allowed
by the worldname validator on AW's website, and also the maximum I've
seen in universe.  This typically gets matched by checking for a word
boundary, but I've conditionally set it to check for ! or @ to match
some special cases for AWI worlds.  New worlds created via the web
dialog aren't allowed to use these special characters., but there
doesn't seem to be any limitation in the actual software to prevent them
from appearing.